pyproject.toml
[project]
name = "omni-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cantools>=41.1.1",
    "python-can>=4.6.1",
]
